+++ src/index.js
@@ -1,0 +1,3 @@
+const Analyzer = require('./analyzer');
+
+module.exports = { Analyzer };